Module nu_ansi_term::ansi

Structs

Infix

Like AnsiString, but only displays the difference between two styles.

Prefix

Like AnsiString, but only displays the style prefix.

Suffix

Like AnsiString, but only displays the style suffix.

Statics

RESET

The code to send to reset all styles and return to Style::default().