nu_ansi_term
Type Alias AnsiString

pub type AnsiString<'a> = AnsiGenericString<'a, str>;

An ANSI String is a string coupled with the Style to display it in a terminal.

Although not technically a string itself, it can be turned into one with the to_string method.

Examples

use nu_ansi_term::AnsiString;
use nu_ansi_term::Color::Red;

let red_string = Red.paint("a red string");
println!("{}", red_string);

use nu_ansi_term::AnsiString;

let plain_string = AnsiString::from("a plain string");

Aliased Type

struct AnsiString<'a> { /* private fields */ }