nu_ansi_term

Struct Rgb

pub struct Rgb {
    pub r: u8,
    pub g: u8,
    pub b: u8,
}

Fields
r: u8

Red

g: u8

Green

b: u8

Blue

Implementations

impl Rgb

pub const fn new(r: u8, g: u8, b: u8) -> Self

Creates a new Rgb color

pub const fn from_hex(hex: u32) -> Self

Creates a new Rgb color with a hex code

pub fn from_hex_string(hex: String) -> Self

pub fn from_f32(r: f32, g: f32, b: f32) -> Self

Creates a new Rgb color with three f32 values

pub const fn gray(x: u8) -> Self

Creates a grayscale Rgb color

pub fn gray_f32(x: f32) -> Self

Creates a grayscale Rgb color with a f32 value

pub fn lerp(&self, other: Self, t: f32) -> Self

Creates a new Rgb color from a [HSL] color Computes the linear interpolation between self and other for t